1_graph_creation/functions/agent1_ingest_source/requirements.txt
functions-framework==3.*
flask==2.*
google-cloud-storage
google-cloud-pubsub
